Format changed to GG's extended jj2lsj notation
Gediminas Gaigalas suggested an extended format for the jj2lsj output
labels, mainly to make room for larger level indices during calculations
of more than 999 levels in a block. This version of rlevels now complies
with that format which should be ok for up to 99 999 levels per block.
Testing has to be done to sync the format over the whole range of grasp
codes.